updates for refactor
