+++ backend/src/utils/middleware.ts
@@ -1,4 +1,9 @@
-import { NextFunction, Request, Response } from 'express';
+import { NextFunction, Response, Request } from 'express';
+import { AuthenticationService } from '../services/authentication-services.js';
+import { UserService } from '../services/user-services.js';
+
+const authService = new AuthenticationService();
+const userService = new UserService();
 
 export const errorMiddleware = (err: any, req: Request, res: Response, next: NextFunction): Response<any> | void => {
     console.error(err);
