+++ frontend/src/pages/Login.tsx
@@ -6,7 +6,6 @@ import { toast } from "react-hot-toast";
 import { useAuth } from "../context/AuthContext";
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { AxiosError } from 'axios';
 
 const Login = () => {
     const auth = useAuth();
@@ -21,9 +20,9 @@ const Login = () => {
             await auth?.login(email, password);
             toast.success("Signed in successfully!", { id: "login" })
         } catch (error: unknown) {
-            if (error instanceof AxiosError) {
+            if (error instanceof Error) {
             console.log(error);
-            toast.error("Sign in failed. " + error.response?.data?.toString() , { id: "login" });
+            toast.error("Sign in failed. " + error.message?.toString() , { id: "login" });
             }
         }
     }
